TIP-706: small refactorization abstract Attribute
