+++ src/Pim/Bundle/CatalogBundle/Elasticsearch/Filter/AbstractFilter.php
@@ -2,11 +2,8 @@
 
 namespace Pim\Bundle\CatalogBundle\Elasticsearch\Filter;
 
-use Akeneo\Component\StorageUtils\Exception\InvalidPropertyException;
 use Pim\Bundle\CatalogBundle\Elasticsearch\SearchQueryBuilder;
-use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Query\Filter\FilterInterface;
-use Pim\Component\Catalog\Validator\AttributeValidatorHelper;
 
 abstract class AbstractFilter implements FilterInterface
 {
@@ -15,12 +12,6 @@ abstract class AbstractFilter implements FilterInterface
 
     /** @var array */
     protected $supportedOperators = [];
-
-    /** @var AttributeValidatorHelper */
-    protected $attrValidatorHelper;
-
-    /** @var string[] */
-    protected $supportedAttributeTypes;
 
     /**
      * {@inheritdoc}
@@ -41,22 +32,6 @@ abstract class AbstractFilter implements FilterInterface
     /**
      * {@inheritdoc}
      */
-    public function getAttributeTypes()
-    {
-        return $this->supportedAttributeTypes;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function supportsAttribute(AttributeInterface $attribute)
-    {
-        return in_array($attribute->getAttributeType(), $this->supportedAttributeTypes);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
     public function setQueryBuilder($searchQueryBuilder)
     {
         if (!$searchQueryBuilder instanceof SearchQueryBuilder) {
@@ -66,45 +41,5 @@ abstract class AbstractFilter implements FilterInterface
         }
 
         $this->searchQueryBuilder = $searchQueryBuilder;
-    }
-
-    /**
-     * Check locale and scope are valid
-     *
-     * @param AttributeInterface $attribute
-     * @param string             $locale
-     * @param string             $scope
-     *
-     * @throws InvalidPropertyException
-     */
-    protected function checkLocaleAndScope(AttributeInterface $attribute, $locale, $scope)
-    {
-        try {
-            $this->attrValidatorHelper->validateLocale($attribute, $locale);
-            $this->attrValidatorHelper->validateScope($attribute, $scope);
-        } catch (\LogicException $e) {
-            throw InvalidPropertyException::expectedFromPreviousException(
-                $attribute->getCode(),
-                static::class,
-                $e
-            );
-        }
-    }
-
-    /**
-     * Calculates the ES path to a product value indexed in ES.
-     *
-     * @param AttributeInterface $attribute
-     * @param string             $locale
-     * @param string             $scope
-     *
-     * @return string
-     */
-    protected function getAttributePath(AttributeInterface $attribute, $locale, $scope)
-    {
-        $locale = (null === $locale) ? '<all_locales>' : $locale;
-        $scope = (null === $scope) ? '<all_channels>' : $scope;
-
-        return 'values.' . $attribute->getCode() . '-' . $attribute->getBackendType() . '.' . $locale . '.' . $scope;
     }
 }
