+++ src/Pim/Bundle/CatalogBundle/Elasticsearch/Filter/TextAreaFilter.php
@@ -16,7 +16,7 @@ use Pim\Component\Catalog\Validator\AttributeValidatorHelper;
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TextAreaFilter extends AbstractFilter implements AttributeFilterInterface
+class TextAreaFilter extends AbstractAttributeFilter implements AttributeFilterInterface
 {
     /**
      * @param AttributeValidatorHelper $attrValidatorHelper
